requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
Flask==2.3.3
fonttools==4.61.0
fsspec==2025.10.0
gunicorn==23.0.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
packaging==25.0
pillow==12.0.0
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1+cpu
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
